Fixing Bfruit and Double Cross
Many Python games requires to KillAudioserver + unset LD_PRELOAD otherwise it crash on the audio start.

Diag : I have successfully tested "noaudiofix":1" in config.json : it works but it break sound on other ports.
The objective is to avoid the automatic load of "LD_PRELOAD=/customer/lib/libpadsp.so".
To do that adding unset LD_PRELOAD is the solution (same as Zelda Picross.port ).